Move CompactionMethod enum to configuration module
